+++ SnowEngine/Source/Component/Component.h
@@ -520,7 +520,7 @@ protected:
 	 *	родительский `tick`.
 	 *	\param delta_sec Длительность тика в секундах.
 	 */
-	virtual void tick(float delta_sec);
+	virtual void tick(double delta_sec);
 
 			/* EVENT METHODS */
